INS-320: Compile moved to preprocessor
